updated whale-api-client with refactor
