+++ packages/whale-api-client/__tests__/api/transaction.test.ts
@@ -1,307 +1,163 @@
 import { MasterNodeRegTestContainer } from '@defichain/testcontainers'
-import { createSignedTxnHex } from '@defichain/testing'
 import { StubWhaleApiClient } from '../stub.client'
 import { StubService } from '../stub.service'
-import { WhaleApiClient, WhaleApiException, WhaleApiValidationException } from '../../src'
+import { WhaleApiClient, WhaleApiException } from '../../src'
+import { JsonRpcClient } from '@defichain/jellyfish-api-jsonrpc'
 
-describe('transaction', () => {
-  let container: MasterNodeRegTestContainer
-  let service: StubService
-  let client: WhaleApiClient
+let container: MasterNodeRegTestContainer
+let service: StubService
+let client: WhaleApiClient
+let rpcClient: JsonRpcClient
+
+beforeAll(async () => {
+  container = new MasterNodeRegTestContainer()
+  service = new StubService(container)
+  client = new StubWhaleApiClient(service)
+
+  await container.start()
+  await container.waitForWalletCoinbaseMaturity()
+  await service.start()
+
+  await container.waitForBlockHeight(101)
+  await service.waitForIndexedHeight(101)
+
+  rpcClient = new JsonRpcClient(await container.getCachedRpcUrl())
+})
+
+afterAll(async () => {
+  try {
+    await service.stop()
+  } finally {
+    await container.stop()
+  }
+})
+
+beforeEach(async () => {
+  await container.waitForWalletBalanceGTE(15)
+})
+
+describe('get', () => {
+  let txid: string
+
+  async function setup (): Promise<void> {
+    const address = await container.getNewAddress()
+    const metadata = {
+      symbol: 'ETH',
+      name: 'ETH',
+      isDAT: true,
+      mintable: true,
+      tradeable: true,
+      collateralAddress: address
+    }
+
+    txid = await container.call('createtoken', [metadata])
+
+    await container.generate(1)
+
+    const height = await container.call('getblockcount')
+
+    await container.generate(1)
+
+    await service.waitForIndexedHeight(height)
+  }
 
   beforeAll(async () => {
-    container = new MasterNodeRegTestContainer()
-    service = new StubService(container)
-    client = new StubWhaleApiClient(service)
-
-    await container.start()
-    await container.waitForReady()
-    await container.waitForWalletCoinbaseMaturity()
-    await service.start()
+    await setup()
   })
 
-  afterAll(async () => {
+  it('should get a single transaction', async () => {
+    const transaction = await client.transactions.get(txid)
+    expect(transaction).toStrictEqual({
+      id: txid,
+      block: {
+        hash: expect.any(String),
+        height: expect.any(Number),
+        time: expect.any(Number),
+        medianTime: expect.any(Number)
+      },
+      txid,
+      hash: txid,
+      version: expect.any(Number),
+      size: expect.any(Number),
+      vSize: expect.any(Number),
+      weight: expect.any(Number),
+      lockTime: expect.any(Number),
+      vinCount: expect.any(Number),
+      voutCount: expect.any(Number)
+    })
+  })
+
+  it('should fail due to non-existent transaction', async () => {
+    expect.assertions(2)
     try {
-      await service.stop()
-    } finally {
-      await container.stop()
+      await client.transactions.get('invalidtransactionid')
+    } catch (err) {
+      expect(err).toBeInstanceOf(WhaleApiException)
+      expect(err.error).toStrictEqual({
+        code: 404,
+        type: 'NotFound',
+        at: expect.any(Number),
+        message: 'Unable to find transaction by id: invalidtransactionid',
+        url: '/v0.0/regtest/transactions/invalidtransactionid'
+      })
     }
   })
+})
 
-  beforeEach(async () => {
-    await container.waitForWalletBalanceGTE(15)
+describe('getVins', () => {
+  it('should return list of vins', async () => {
+    const blockHash = await container.call('getblockhash', [100])
+    const block = await rpcClient.blockchain.getBlock(blockHash, 2)
+
+    const txid = block.tx[0].txid
+    const vins = await client.transactions.getVins(txid)
+
+    expect(vins.length).toBeGreaterThanOrEqual(1)
   })
 
-  describe('test', () => {
-    it('should accept valid txn', async () => {
-      const hex = await createSignedTxnHex(container, 10, 9.9999)
-      await client.transactions.send({
-        hex: hex
-      })
-    })
+  it('should return list of vins when next is out of range', async () => {
+    const blockHash = await container.call('getblockhash', [100])
+    const block = await rpcClient.blockchain.getBlock(blockHash, 2)
 
-    it('should accept valid txn with given maxFeeRate', async () => {
-      const hex = await createSignedTxnHex(container, 10, 9.995)
-      await client.transactions.test({
-        hex: hex, maxFeeRate: 0.05
-      })
-    })
+    const txid = block.tx[0].txid
+    const vins = await client.transactions.getVins(txid, 30, '100')
 
-    it('should reject due to invalid txn', async () => {
-      expect.assertions(2)
-      try {
-        await client.transactions.test({ hex: '0400000100881133bb11aa00cc' })
-      } catch (err) {
-        expect(err).toBeInstanceOf(WhaleApiException)
-        expect(err.error).toStrictEqual({
-          code: 400,
-          type: 'BadRequest',
-          message: 'Transaction decode failed',
-          at: expect.any(Number),
-          url: '/v0.0/regtest/transactions/test'
-        })
-      }
-    })
-
-    it('should reject due to high fees', async () => {
-      const hex = await createSignedTxnHex(container, 10, 9)
-      expect.assertions(2)
-      try {
-        await client.transactions.test({
-          hex: hex, maxFeeRate: 1
-        })
-      } catch (err) {
-        expect(err).toBeInstanceOf(WhaleApiException)
-        expect(err.error).toStrictEqual({
-          code: 400,
-          type: 'BadRequest',
-          at: expect.any(Number),
-          message: 'Transaction is not allowed to be inserted',
-          url: '/v0.0/regtest/transactions/test'
-        })
-      }
-    })
+    expect(vins.length).toBeGreaterThanOrEqual(1)
   })
 
-  describe('send', () => {
-    it('should send valid txn', async () => {
-      const hex = await createSignedTxnHex(container, 10, 9.9999)
-      const txid = await client.transactions.send({
-        hex: hex
-      })
-      expect(txid.length).toStrictEqual(64)
+  it('should return empty page if txid is not valid', async () => {
+    const vins = await client.transactions.getVins('9d87a6b6b77323b6dab9d8971fff0bc7a6c341639ebae39891024f4800528532', 30)
 
-      await container.generate(1)
-      const out = await container.call('gettxout', [txid, 0])
-      expect(out.value).toStrictEqual(9.9999)
-    })
+    expect(vins.length).toStrictEqual(0)
+    expect(vins.hasNext).toStrictEqual(false)
+  })
+})
 
-    it('should send valid txn with given maxFeeRate', async () => {
-      const hex = await createSignedTxnHex(container, 10, 9.995)
-      const txid = await client.transactions.send({
-        hex: hex, maxFeeRate: 0.05
-      })
-      expect(txid.length).toStrictEqual(64)
+describe('getVouts', () => {
+  it('should return list of vouts', async () => {
+    const blockHash = await container.call('getblockhash', [100])
+    const block = await rpcClient.blockchain.getBlock(blockHash, 2)
 
-      await container.generate(1)
-      const out = await container.call('gettxout', [txid, 0])
-      expect(out.value).toStrictEqual(9.995)
-    })
+    const txid = block.tx[0].txid
+    const vouts = await client.transactions.getVouts(txid)
 
-    it('should fail due to invalid txn', async () => {
-      expect.assertions(2)
-      try {
-        await client.transactions.send({
-          hex: '0400000100881133bb11aa00cc'
-        })
-      } catch (err) {
-        expect(err).toBeInstanceOf(WhaleApiException)
-        expect(err.error).toStrictEqual({
-          code: 400,
-          type: 'BadRequest',
-          at: expect.any(Number),
-          url: '/v0.0/regtest/transactions',
-          message: 'Transaction decode failed'
-        })
-      }
-    })
-
-    it('should fail due to high fees', async () => {
-      const hex = await createSignedTxnHex(container, 10, 9)
-
-      expect.assertions(2)
-      try {
-        await client.transactions.send({
-          hex: hex, maxFeeRate: 1
-        })
-      } catch (err) {
-        expect(err).toBeInstanceOf(WhaleApiException)
-        expect(err.error).toStrictEqual({
-          code: 400,
-          type: 'BadRequest',
-          at: expect.any(Number),
-          url: '/v0.0/regtest/transactions',
-          message: 'Absurdly high fee'
-        })
-      }
-    })
-
-    it('should fail validation (empty hex)', async () => {
-      expect.assertions(3)
-      try {
-        await client.transactions.send({
-          hex: ''
-        })
-        expect('must fail').toBeUndefined()
-      } catch (err) {
-        expect(err).toBeInstanceOf(WhaleApiValidationException)
-        expect(err.message).toStrictEqual('422 - ValidationError (/v0.0/regtest/transactions)')
-        expect(err.properties).toStrictEqual([{
-          constraints: [
-            'hex must be a hexadecimal number',
-            'hex should not be empty'
-          ],
-          property: 'hex',
-          value: ''
-        }])
-      }
-    })
-
-    it('should fail validation (not hex)', async () => {
-      expect.assertions(3)
-      try {
-        await client.transactions.send({
-          hex: 'fuxingloh'
-        })
-        expect('must fail').toBeUndefined()
-      } catch (err) {
-        expect(err).toBeInstanceOf(WhaleApiValidationException)
-        expect(err.message).toStrictEqual('422 - ValidationError (/v0.0/regtest/transactions)')
-        expect(err.properties).toStrictEqual([{
-          constraints: [
-            'hex must be a hexadecimal number'
-          ],
-          property: 'hex',
-          value: 'fuxingloh'
-        }])
-      }
-    })
-
-    it('should fail validation (negative fee)', async () => {
-      expect.assertions(3)
-      try {
-        await client.transactions.send({
-          hex: '00', maxFeeRate: -1.5
-        })
-        expect('must fail').toBeUndefined()
-      } catch (err) {
-        expect(err).toBeInstanceOf(WhaleApiValidationException)
-        expect(err.message).toStrictEqual('422 - ValidationError (/v0.0/regtest/transactions)')
-        expect(err.properties).toStrictEqual([{
-          constraints: [
-            'maxFeeRate must not be less than 0'
-          ],
-          property: 'maxFeeRate',
-          value: -1.5
-        }])
-      }
-    })
-
-    it('should fail validation (not number fee)', async () => {
-      expect.assertions(3)
-      try {
-        await client.transactions.send({
-          // @ts-expect-error
-          hex: '00', maxFeeRate: 'abc'
-        })
-        expect('must fail').toBeUndefined()
-      } catch (err) {
-        expect(err).toBeInstanceOf(WhaleApiValidationException)
-        expect(err.message).toStrictEqual('422 - ValidationError (/v0.0/regtest/transactions)')
-        expect(err.properties).toStrictEqual([{
-          constraints: [
-            'maxFeeRate must not be less than 0',
-            'maxFeeRate must be a number conforming to the specified constraints'
-          ],
-          property: 'maxFeeRate',
-          value: 'abc'
-        }])
-      }
-    })
+    expect(vouts.length).toBeGreaterThanOrEqual(1)
   })
 
-  describe('estimateFee', () => {
-    it('should be fixed fee of 0.00005000 when there are no transactions', async () => {
-      const feeRate = await client.transactions.estimateFee(10)
-      expect(feeRate).toStrictEqual(0.00005000)
-    })
+  it('should return list of vouts when next is out of range', async () => {
+    const blockHash = await container.call('getblockhash', [100])
+    const block = await rpcClient.blockchain.getBlock(blockHash, 2)
+
+    const txid = block.tx[0].txid
+    const vouts = await client.transactions.getVouts(txid, 30, '100')
+
+    expect(vouts.length).toBeGreaterThanOrEqual(1)
   })
 
-  describe('get', () => {
-    let txid: string
+  it('should return empty page if txid is not valid', async () => {
+    const vouts = await client.transactions.getVouts('9d87a6b6b77323b6dab9d8971fff0bc7a6c341639ebae39891024f4800528532', 30)
 
-    async function setup (): Promise<void> {
-      const address = await container.getNewAddress()
-      const metadata = {
-        symbol: 'ETH',
-        name: 'ETH',
-        isDAT: true,
-        mintable: true,
-        tradeable: true,
-        collateralAddress: address
-      }
-
-      txid = await container.call('createtoken', [metadata])
-
-      await container.generate(1)
-
-      const height = await container.call('getblockcount')
-
-      await container.generate(1)
-
-      await service.waitForIndexedHeight(height)
-    }
-
-    beforeAll(async () => {
-      await setup()
-    })
-
-    it('should get a single transaction', async () => {
-      const transaction = await client.transactions.get(txid)
-      expect(transaction).toStrictEqual({
-        id: txid,
-        block: {
-          hash: expect.any(String),
-          height: expect.any(Number),
-          time: expect.any(Number),
-          medianTime: expect.any(Number)
-        },
-        txid,
-        hash: txid,
-        version: expect.any(Number),
-        size: expect.any(Number),
-        vSize: expect.any(Number),
-        weight: expect.any(Number),
-        lockTime: expect.any(Number),
-        vinCount: expect.any(Number),
-        voutCount: expect.any(Number)
-      })
-    })
-
-    it('should fail due to non-existent transaction', async () => {
-      expect.assertions(2)
-      try {
-        await client.transactions.get('invalidtransactionid')
-      } catch (err) {
-        expect(err).toBeInstanceOf(WhaleApiException)
-        expect(err.error).toStrictEqual({
-          code: 404,
-          type: 'NotFound',
-          at: expect.any(Number),
-          message: 'Unable to find transaction by id: invalidtransactionid',
-          url: '/v0.0/regtest/transactions/invalidtransactionid'
-        })
-      }
-    })
+    expect(vouts.length).toStrictEqual(0)
+    expect(vouts.hasNext).toStrictEqual(false)
   })
 })
