+++ packages/whale-api-client/src/whale.api.client.ts
@@ -14,6 +14,8 @@ import { Blocks } from './api/blocks'
 import { Oracles } from './api/oracles'
 import { Prices } from './api/prices'
 import { Stats } from './api/stats'
+import { Rawtx } from './api/rawtx'
+import { Fee } from './api/fee'
 
 /**
  * WhaleApiClient Options
@@ -70,6 +72,8 @@ export class WhaleApiClient {
   public readonly oracles = new Oracles(this)
   public readonly prices = new Prices(this)
   public readonly stats = new Stats(this)
+  public readonly rawtx = new Rawtx(this)
+  public readonly fee = new Fee(this)
 
   constructor (
     protected readonly options: WhaleApiClientOptions
